refactor(heap): Heap内部ストレージをVec<u64>からVec<u8>に移行しバイトアドレッシングに統一
将来のサブワード型(U8, I32)サポートや文字列効率化の基盤として、Heap.memoryを
Vec<u64>(ワードアドレッシング)からVec<u8>(バイトアドレッシング)に移行。
GcRefのオフセットもワード単位からバイト単位に変更し、JITコンパイラの
ヒープアクセスコード生成を対応させた。

- heap.rs: read_u64/write_u64ヘルパー追加、全内部アドレッシングをバイト単位に変更
- marshal.rs: JitCallContext.heap_baseを*const u64から*const u8に変更
- compiler_microop_x86_64.rs: shl 3パターンを除去しバイトオフセット直接計算に変更
- compiler_microop.rs: aarch64 JITも同様にバイトアドレッシングに対応

Closes #264, Closes #265

+++ src/jit/marshal.rs
@@ -185,8 +185,8 @@ pub struct JitCallContext {
     pub hostcall_helper:
         unsafe extern "C" fn(*mut JitCallContext, u64, u64, *const JitValue) -> JitReturn,
     /// Pointer to heap memory base (for direct heap access from JIT code)
-    /// This points to the first element of the heap's memory Vec<u64>.
-    pub heap_base: *const u64,
+    /// This points to the first byte of the heap's memory Vec<u8>.
+    pub heap_base: *const u8,
     /// Pointer to string constant cache (for direct access from JIT code)
     /// This points to the first element of VM's string_cache Vec<Option<GcRef>>.
     /// Each entry is 16 bytes: Option<GcRef> where GcRef is 8 bytes (usize).
